refactor: move read functions out of the cli into a helper file
